requirements.txt
# ETH Options Tick Data Collector - Python Dependencies
aiohttp==3.10.11
asyncpg==0.29.0
pyyaml==6.0.2
requests==2.32.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
